requirements.txt
matplotlib==3.10.0
nano-vectordb==0.0.4.3
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.65.2
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
tiktoken==0.9.0
tqdm==4.67.1

# API Dependencies
fastapi==0.115.12
pydantic==2.10.6
uvicorn==0.34.1
aiolimiter~=1.2.1